Release resourses after stop recording

+++ views/js/pciCreator/dev/audioRecordingInteraction/runtime/js/providers/webAudio.js
@@ -172,14 +172,11 @@ define([
              * Close the audio context and destroy created assets
              */
             destroy: function destroy() {
-                recorderWorker.terminate();
-                recorderWorker = null;
-
-                if (getAudioContext()) {
-                    return getAudioContext().close().then(function() {
-                        audioNodes = {};
-                    });
+                if (recorderWorker) {
+                    recorderWorker.terminate();
+                    recorderWorker = null;
                 }
+                audioNodes = {};
             }
         };
